opdate to use std::string

+++ myctrl_tidal2.h
@@ -247,7 +247,7 @@ class tidal_class {
     void set_textureloaded(bool set);
 
     int save_music_oversigt_playlists(char *playlistfilename,int tidalknapnr,char *cover_path,char *playlstid,char *artistname);
-    bool delete_record_in_view(int tidalknapnr);
+    bool delete_record_in_view(long tidalknapnr);
     int get_users_playlist_plus_favorite(bool cleandb);
 
     // int tidal_play_playlist(char *playlist_song,int tidalknapnr,bool now);
